finally get on popup delegate working
