+++ Source/CesiumRuntime/Private/MyScreenCreditsBase.h
@@ -3,6 +3,7 @@
 #pragma once
 
 #include "Blueprint/UserWidget.h"
+#include "Components/RichTextBlockDecorator.h"
 #include "CoreMinimal.h"
 #include "Interfaces/IHttpRequest.h"
 #include <memory>
@@ -13,6 +14,8 @@ namespace Cesium3DTilesSelection {
 class CreditSystem;
 struct Credit;
 } // namespace Cesium3DTilesSelection
+
+DECLARE_DELEGATE(FOnPopupClicked)
 
 UCLASS()
 class UMyScreenCreditsBase : public UUserWidget {
@@ -26,6 +29,8 @@ private:
   UFUNCTION(BlueprintCallable)
   void Update();
 
+  void OnPopupClicked();
+
   UPROPERTY(meta = (BindWidget))
   class URichTextBlock* RichTextBlock_127;
 
@@ -38,6 +43,26 @@ private:
   std::shared_ptr<Cesium3DTilesSelection::CreditSystem> _pCreditSystem;
   size_t _lastCreditsCount;
   TMap<const Cesium3DTilesSelection::Credit*, FString> _creditToRTF;
-  UMyRichTextBlockDecorator* _imageDecorator;
+  class UMyRichTextBlockDecorator* _imageDecorator;
   FString _output;
+  bool _showPopup;
+};
+
+UCLASS()
+class UMyRichTextBlockDecorator : public URichTextBlockDecorator {
+  GENERATED_BODY()
+
+public:
+  UMyRichTextBlockDecorator(const FObjectInitializer& ObjectInitializer);
+
+  virtual TSharedPtr<ITextDecorator>
+  CreateDecorator(URichTextBlock* InOwner) override;
+
+  virtual const FSlateBrush* FindImageBrush(int32 id);
+
+private:
+  TArray<FSlateDynamicImageBrush*> _textureResources;
+  FOnPopupClicked EventHandler;
+  friend class UMyScreenCreditsBase;
+  friend class FRichInlineImage;
 };
